+++ crates/pipeline_manager/migrations/V10__remove_program_status_none.sql
@@ -1,0 +1,3 @@
+-- Set all programs to be automatically compiled. Removes 'none'
+-- as a valid status field in the database.
+UPDATE program SET status = 'pending' WHERE status = 'none';